requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.7.0
matplotlib>=3.4.0
requests>=2.26.0
